Moved endpoints to a separate controller
